requirements.txt
asgiref==3.2.5
astroid==2.3.3
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.4
django-filter==2.2.0
djangorestframework==3.11.0
drf-generators==0.5.0
drf-yasg==1.17.1
idna==2.9
importlib-metadata==1.5.0
inflection==0.3.1
isort==4.3.21
itypes==1.1.0
Jinja2==2.11.1
jsonschema==3.2.0
lazy-object-proxy==1.4.3
Markdown==3.2.1
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==1.4.6
packaging==20.3
pylint==2.4.4
pylint-django==2.0.15
pylint-plugin-utils==0.6
PyMySQL==0.9.3
pyparsing==2.4.6
pyrsistent==0.15.7
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.14.0
sqlparse==0.3.1
swagger-spec-validator==2.5.0
uritemplate==3.0.1
urllib3==1.25.8
wrapt==1.11.2
zipp==3.1.0
